chore: Move cellDescription into its own file
